requirements.txt
psutil
datetime
mysql-connector-python
python-dotenv
Flask
mysql